Cipher refactor
