+++ src/emrtd/pace.rs
@@ -1,6 +1,6 @@
 use {
     super::Emrtd,
-    crate::{asn1::public_key_info::EcParameters, emrtd::secure_messaging::aes::kdf_128},
+    crate::{asn1::public_key_info::EcParameters, crypto::cipher::aes::kdf_128},
     anyhow::Result,
     rand::{CryptoRng, RngCore},
     sha1::{Digest, Sha1},
@@ -12,7 +12,7 @@ impl Emrtd {
     pub fn pace(&mut self, _rng: impl CryptoRng + RngCore, mrz: &str) -> Result<()> {
         // Derive symmetric key K_pi
         let k = k_from_mrz(mrz);
-        let _k_pi = kdf_128(&k[..], KDF_PACE);
+        let _k_pi = kdf_128(&k[..], KDF_PACE)?;
 
         // Send MSE:Set AT.
 
@@ -41,7 +41,7 @@ pub fn standardized_parameters(id: u64) -> Option<EcParameters> {
 
 #[cfg(test)]
 mod tests {
-    use {super::*, crate::emrtd::secure_messaging::aes::kdf_128, hex_literal::hex};
+    use {super::*, crate::crypto::cipher::aes::kdf_128, hex_literal::hex};
 
     // ICAO 9303-11, Appendix G
     #[test]
@@ -49,7 +49,7 @@ mod tests {
         let mrz = "T22000129364081251010318";
         let k = k_from_mrz(mrz);
         assert_eq!(k, hex!("7E2D2A41 C74EA0B3 8CD36F86 3939BFA8 E9032AAD"));
-        let k_pi = kdf_128(&k[..], 3);
+        let k_pi = kdf_128(&k[..], 3).unwrap();
         assert_eq!(k_pi, hex!("89DED1B2 6624EC1E 634C1989 302849DD"));
 
         // let pace_info = PaceInfo::from_der(&hex!("3012060A 04007F00 07020204
